rename SparseValuePath to SimpleSmtProof
